Fix CI: move sidebar unit tests to reason-editor-sidebar, wire up its test runner
The sidebar-extraction commit moved the source files under
packages/reason-editor/src/search/{OutlineView,useActiveHeading,relatedDocuments}
into the new package but missed the corresponding test files under
packages/reason-editor/test/search/, which broke reason-editor's CI job
with unresolved "@/search/..." imports.

- Move those three test files into packages/reason-editor-sidebar/test/,
  updating their imports to the new package's module layout.
- Add a vitest.config.ts (jsdom environment, @ alias, coverage scope) and
  a test/test:coverage script + vitest/jsdom/coverage-v8 devDependencies to
  packages/reason-editor-sidebar, mirroring reason-editor's own setup.
- Fix test/search/searchDocuments.test.ts's Document import (that test and
  its source stay in reason-editor) to pull the type from
  reason-editor-sidebar instead of the now-deleted local path.
- Register the new package in the root vitest.config.ts project list and
  in the Coverage workflow's matrix so its own tests actually run in CI.
