+++ packages/reason-editor-sidebar/package.json
@@ -35,6 +35,8 @@
     "dev": "vite build --watch",
     "build": "vite build && tsc --project tsconfig.build.json",
     "type-check": "tsc --noEmit",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
     "prepublishOnly": "bun run build"
   },
   "keywords": [
@@ -68,10 +70,13 @@
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^6.0.2",
+    "@vitest/coverage-v8": "^4.0.18",
+    "jsdom": "^28.1.0",
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
     "typescript": "^6.0.3",
-    "vite": "^8.0.16"
+    "vite": "^8.0.16",
+    "vitest": "^4.0.18"
   },
   "license": "rights.institute/PROSPER"
 }
