+++ packages/reason-editor-sidebar/test/outline/OutlineView.test.tsx
@@ -9,9 +9,9 @@ import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import { OutlineView, computeScrollIntoViewOffset } from '@/search/OutlineView';
-import type { ActiveHeadingEditorHandle } from '@/search/useActiveHeading';
-import type { TocEntry } from '@/app-types/toc';
+import { OutlineView, computeScrollIntoViewOffset } from '@/outline/OutlineView';
+import type { ActiveHeadingEditorHandle } from '@/outline/useActiveHeading';
+import type { TocEntry } from '@/types/toc';
 
 (globalThis as any).IS_REACT_ACT_ENVIRONMENT = true;
