+++ packages/reason-editor-sidebar/tsconfig.json
@@ -12,8 +12,11 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "types": []
+    "types": [],
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   },
-  "include": ["src", "vite.config.ts"],
+  "include": ["src", "test", "vite.config.ts", "vitest.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
