+++ vitest.config.ts
@@ -30,6 +30,7 @@ export default defineConfig({
       'packages/qwksearch-mcp-server',
       'packages/react-weather-forecast',
       'packages/reason-editor',
+      'packages/reason-editor-sidebar',
       'packages/render-url-to-html/scraper-jsdom',
       'packages/render-url-to-html/scraper-puppeteer',
       'packages/research-agent-ui',
